renaming decision support providers
