+++ cmd/demo/demo.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hexa-org/policy-orchestrator/cmd/demo/azuresupport"
 	"github.com/hexa-org/policy-orchestrator/cmd/demo/googlesupport"
 	"github.com/hexa-org/policy-orchestrator/pkg/decisionsupport"
-	"github.com/hexa-org/policy-orchestrator/pkg/decisionsupport/providers"
+	"github.com/hexa-org/policy-orchestrator/pkg/decisionsupport/decisionproviders"
 	"github.com/hexa-org/policy-orchestrator/pkg/websupport"
 	"log"
 	"net"
@@ -27,7 +27,7 @@ func App(session *sessions.CookieStore, amazonConfig amazonsupport.AmazonCognito
 	googleSupport := googlesupport.NewGoogleSupport(session)
 	amazonSupport := amazonsupport.NewAmazonSupport(client, amazonConfig, amazonsupport.AmazonCognitoClaimsParser{}, session)
 	azureSupport := azuresupport.NewAzureSupport(session)
-	provider := providers.OpaDecisionProvider{Client: client, Url: opaUrl, Principals: []interface{}{"allusers", "allauthenticatedusers", "sales@hexaindustries.io"}}
+	provider := decisionproviders.OpaDecisionProvider{Client: client, Url: opaUrl, Principals: []interface{}{"allusers", "allauthenticatedusers", "sales@hexaindustries.io"}}
 	opaSupport := decisionsupport.DecisionSupport{Provider: provider, Unauthorized: basic.unauthorized, Skip: []string{"/health", "/metrics", "/styles", "/images", "/bundle", "/favicon.ico"}}
 	server := websupport.Create(addr, basic.loadHandlers(), websupport.Options{ResourceDirectory: resourcesDirectory})
 	router := server.Handler.(*mux.Router)
