+++ pkg/decisionsupport/decisionproviders/mock_provider.go
@@ -1,4 +1,4 @@
-package providers
+package decisionproviders
 
 import (
 	"github.com/stretchr/testify/mock"
